pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "decision-debate"
version = "0.1.0"
description = "Three-agent decision support tool: PRO / CON / JUDGE powered by Gemini and LangGraph."
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.8.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0",
    "typer>=0.12.0",
    "rich>=13.0",
]

[project.scripts]
agent-debate = "agent_debate.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_debate*"]
